finished events, membership, attendance endpoints and added auth middleware

+++ backend/db/migrations/20231128160717-create-event.js
@@ -48,7 +48,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.NUMERIC,
         allowNull: false
       },
       startDate: {
